refactor: HintItem으로 Strike, Ball, Nothing 추상화
겹치는 부분들이 많아 abstract class로 추상화함.
