+++ src/main/java/baseball/domain/Hint/HintItem.java
@@ -1,0 +1,15 @@
+package baseball.domain.Hint;
+
+public abstract class HintItem {
+
+    private int count;
+
+    public int getCount() {
+        return count;
+    }
+
+    public void setCount(int count) {
+        this.count = count;
+    }
+
+}
